requirements.txt
pygame
python-chess 